Figured out how to get attachments
Same as title.
Also redid the command files to be made with SlashCommandBuilder class instead of using the data {} thing.
Currently working on saving images to a folder, and then being able to call those images. If all goes well, I might be able to add extra info for each image.